pyproject.toml
[tool.poetry]
name = "valerie-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.9"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
ptbcontrib = {git = "https://github.com/python-telegram-bot/ptbcontrib.git", rev = "main"}
setuptools = "^75.6.0"
openai = "^1.58.1"
langchain-openai = "^0.2.14"
python-json-logger = "^3.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
mocker = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
